requirements.txt
argon2-cffi==16.3.0
bcrypt==3.1.3
boto3==1.4.4
botocore==1.5.84
cffi==1.10.0
Django==1.11.3
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-storages==1.6.3
docutils==0.13.1
gunicorn==19.7.1
jmespath==0.9.3
markdown2==2.3.4
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
